double to float

+++ libs/rcore/rcolor.cpp
@@ -1,6 +1,6 @@
 #include "rcolor.h"
 
-raylib::Color TRNSP(raylib::Color color, double transparent) {
+raylib::Color TRNSP(raylib::Color color, float transparent) {
     color.a = round(color.a * transparent);
     return color;
 }
